feat: Restrict dashboard access to explicit permission
The dashboard can now only be accessed if the user has the explicit 'dashboard' permission in their roles. Previously, access was granted if the user had 'novo_financeiro' permission. This change clarifies and enforces dashboard access control.

Additionally, the 'dashboard' module has been added to the list of available modules in User Management and to the default 'gerente' profile.